fixed an issue when using GPIO feature
When using the GPIO feature, and hence using pins with values above 512, there is an issue with variable ranges.
Using ESP32, compiler warns that the pin value is implicitly truncated for some of the pins. This seems to happen only when the #define of the pin value is used directly in the digitalWriteEx() function call.

The problem seem to be resolved by setting a uint16_t variable to the #define value, before calling the digitalWriteEx().
The issue seem to involve, at minimum, the following pins: ADDON_RESET_PIN, ADDON_GPIO0_PIN, MOUNT_STATUS_LED_PIN, PARK_STATUS_PIN, STATUS_BUZZER_PIN.
The problem also affect the PPS_SENSE_PIN. However since this can not be used with GPIO extender, it was not included in the fix.